Moved to new test module for apply mutation
